+++ boatswain/__init__.py
@@ -2,11 +2,14 @@
 from .boatswain import Boatswain
 from .util import find_dependencies, extract_id, extract_step
 from .cli import argparser
+from .display import Node, Tree
 
 __all__ = [
     "Boatswain",
     "argparser",
     "find_dependencies",
     "extract_id",
-    "extract_step"
+    "extract_step",
+    "Node",
+    "Tree"
 ]