+++ test/test_display.py
@@ -1,0 +1,24 @@
+"""
+    Boatswain test files
+
+    Currently tests only very basic things
+"""
+import pytest
+import docker
+from boatswain import Boatswain, Tree
+
+
+def test_extract(bsfile):
+    tree = Tree()
+    root = tree.extract_tree(bsfile)
+    assert len(root.children) > 0
+
+
+def test_find_child(bsfile):
+    tree = Tree()
+    root = tree.extract_tree(bsfile)
+    assert root.find_child('image3:pytest')
+
+def test_print(bsfile):
+    tree = Tree()
+    tree.print_boatswain_tree(bsfile)